backend/requirements.txt
beautifulsoup4==4.14.0
chromadb==1.2.1
fastapi==0.115.6
google-ai-generativelanguage==0.7.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.78
langchain-deepseek==0.1.3
langchain-google-genai==2.1.12
langchain-tavily==0.2.12
langchain-text-splitters==0.3.8
langgraph==1.0.0
numpy==2.2.4
pydantic==2.11.3
python-dotenv==1.1.0
rank-bm25==0.2.2
requests==2.32.3
rich==14.1.0
sentence-transformers==5.1.1
uvicorn==0.34.0
